Renamed netbox_lifecycle/templatetags/filters.py to netbox_lifecycle_filters.py and updated all {% load filters %} references
