+++ netbox_lifecycle/templates/netbox_lifecycle/hardwarelifecycle.html
@@ -2,7 +2,7 @@
 {% load buttons %}
 {% load custom_links %}
 {% load helpers %}
-{% load filters %}
+{% load netbox_lifecycle_filters %}
 {% load perms %}
 {% load plugins %}
 {% load tabs %}
