+++ netbox_lifecycle/templates/netbox_lifecycle/htmx/contract_list.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load filters %}
+{% load netbox_lifecycle_filters %}
 
 <table class="table table-sm table-hover mb-0 mt-1">
   <thead class="table-light">
