+++ netbox_lifecycle/templates/netbox_lifecycle/htmx/device_contracts.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load filters %}
+{% load netbox_lifecycle_filters %}
 
 <div class="card">
   <h5 class="card-header">{% trans "Support Contracts" %}</h5>
