+++ netbox_lifecycle/templates/netbox_lifecycle/inc/hardware_lifecycle_info.html
@@ -1,5 +1,5 @@
 
-{% load filters %}
+{% load netbox_lifecycle_filters %}
 {% load helpers %}
 {# renders panel on object with lifecycle info assigned to it #}
