+++ netbox_lifecycle/templates/netbox_lifecycle/inc/support_contract_info.html
@@ -1,5 +1,5 @@
 
-{% load filters %}
+{% load netbox_lifecycle_filters %}
 {% load helpers %}
 {# renders panel on object (device) with support contract info assigned to it #}
